Separate the Main component
