+++ modules/core/client-react/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as FieldAdapter } from './FieldAdapter';
 export { default as clientOnly } from './clientOnly';
 export { default as RedBox } from './RedBox';
+export { default as Main } from './Main';
